Eliminate internal TranslationTableOpcode values
These values were stored in the TranslationTableRules for braille
indicators to fully identify them because rules to define braille
indicators sometimes require other keywords beside the opcode. E.g. in
the rule

    endemphphrase italic after 56

"endemphphrase" is the opcode, but the rule was stored with value
CTO_EndEmph1PhraseBeforeRule (not CTO_EndEmphPhrase). This is
potentially useful for back-translation, but the current
back-translation code does not make use of it.

In order to simplify the code we're removing the internal
opcodes. When we want to improve back-translation, we'll come up with
a better way to store the necessary information in the
TranslationTableRule.

Fixes https://github.com/liblouis/liblouis/issues/1170